frontend react and tailwind setup

+++ backend/controllers/product.controller.js
@@ -52,11 +52,15 @@ export const putProduct = async (req, res) => {
 export const deleteProduct = async (req,res) => {
   const {id} = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({ success: false, message: 'Invalid product Id' });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({success:true, message: "product deleted" });
   } catch (error) {
     console.error("Error in delete product:", error.message);
-    res.status(404).json({ success: false, message: "product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
